+++ src/shared/messages.ts
@@ -508,6 +508,7 @@ export const ARDUINO_MESSAGES = {
   CLI_UNINSTALL_LIBRARY: 'cli_UnInstallLibrary',
   CLI_LIBRARY_INSTALLED: 'cli_libraryInstalled',
   CLI_BOARD_EXAMPLES: 'cli_boardExamples',
+  CLI_BURN_BOOTLOADER: 'cli_burnBootloader',
   CLI_LIBRARY_INSTALLATION_ERROR: 'cli_libraryInstallationError',
   CLI_GET_CONFIG: 'cli_getConfig',
   CLI_CONFIG_REMOVE_ADDITIONAL_URL: 'cli_congifRemoveAdditionnalUrl',